Reverting firewalld configuration
As the Fedora 28 doesn't have Python 2 bindings to firewalld and
ansible needs them by default to manage firewalld I'm removing
the firewalld configuration from the deployment process of the
runner.

Please note that as the part of FreeIPA team you should not worry
about it as our infrastructure has external firewall which manages
security rules.