remote_storage: switch from Fedora infra to AWS S3
Currently we are updating test results in Fedora infra where the results are
directly served as part of web server directory listing capabilities. This is
not case of AWS S3.
In order to simulate same environment we are generating "index.html" using
Jinja2 template and putting it into every directory. Then we upload every file
and directory using boto3 library. Before upload we also set correct
"ContentEncoding" and "ContentType" so files (.gz, .png) are directly served.
At the end we create root jobs index to list all jobs in the bucket.