root_index: switch from object metadata to AWS DynamoDB
We are using jobs root index for a convenience of easily
finding and filtering jobs results, however getting data from
object metadata is very slow and time consuming which can result
in task being timed-out. DynamoDB as part of Amazon AWS stack is
more suitable for this task.

As this functionality is not part of main PRCI job a run should not
fail on this.